backend/requirements.txt
Flask
Flask-Cors
Pillow
requests
sqlite3-binary
